Relieve number constraints on Float64 and Int64

+++ src/indsnetworkmap.jl
@@ -95,11 +95,11 @@ function vertex_digit(inm::IndsNetworkMap, v)
   return digit(inm, only(inds(inm, v)))
 end
 
-function dimension_vertices(inm::IndsNetworkMap, dimension::Int64)
+function dimension_vertices(inm::IndsNetworkMap, dimension::Int)
   return filter(v -> vertex_dimension(inm, v) == dimension, vertices(inm))
 end
 
-function vertex(inm::IndsNetworkMap, dimension::Int64, digit::Int64)
+function vertex(inm::IndsNetworkMap, dimension::Int, digit::Int)
   index = ind(inm, dimension, digit)
   return only(filter(v -> index ∈ inm[v], vertices(inm)))
 end
